test: remove useless executor tests and move everything into a single mod file
